+++ src/Control/Eff/Concurrent/Api/Client.hs
@@ -18,7 +18,7 @@ where
 import           Control.Eff
 import           Control.Eff.Reader.Strict
 import           Control.Eff.Concurrent.Api
-import           Control.Eff.Concurrent.Api.Internal
+import           Control.Eff.Concurrent.Api.Request
 import           Control.Eff.Concurrent.Process
 import           Data.Typeable                  ( Typeable )
 import           Control.DeepSeq